Added chat with bing message command

+++ events/discord/messageCreate.js
@@ -20,7 +20,9 @@ module.exports = {
         if (!command && client.config.prefix.length == 0)
             return;
 
-        setTimeout(() => message.delete(), 15000);
+        if (!!command.deleteInvocation)
+            setTimeout(() => message.delete(), 15000);
+
         try {
             if (command.inVoice) {
                 if (message.member.voice.channel === undefined)
@@ -68,8 +70,14 @@ module.exports = {
                 react = [],
                 handler = null,
                 actionRows = null,
-                resetTimeout = false
+                resetTimeout = false,
+                reply = true,
+                deleteResponse = false,
+                content = null
             } = await command.execute(client, queue, message, params);
+
+            if(!!content)
+                return await message.channel.send(content);
 
             const embed = new EmbedBuilder()
                 .setTitle(title)
@@ -81,23 +89,44 @@ module.exports = {
                 .setTimestamp()
                 .setFooter({ text: client.user.username, iconURL: client.botURL });
 
-            await message.reply({ embeds: [embed], components: actionRows }).then(msg => {
-                if (react.length > 0) {
-                    const newHandler = (reaction, user) => handler(reaction, user, msg, queue, client, newHandler);
-                    react.forEach(emoji => msg.react(emoji))
-                    client.on('messageReactionAdd', newHandler);
-                }
-                else {
-                    if (resetTimeout) {
-                        const timeout = setTimeout(() => msg.delete(), 20000);
-                        client.timeouts[msg.id] = {
-                            timeout,
-                            msg
-                        };
+            if (reply)
+                await message.reply({ embeds: [embed], components: actionRows }).then(msg => {
+                    if (react.length > 0) {
+                        const newHandler = (reaction, user) => handler(reaction, user, msg, queue, client, newHandler);
+                        react.forEach(emoji => msg.react(emoji))
+                        client.on('messageReactionAdd', newHandler);
                     }
-                    setTimeout(() => msg.delete(), 15000);
-                }
-            });
+                    else {
+                        if (resetTimeout) {
+                            const timeout = setTimeout(() => msg.delete(), 20000);
+                            client.timeouts[msg.id] = {
+                                timeout,
+                                msg
+                            };
+                        }
+                        if (deleteResponse)
+                            setTimeout(() => msg.delete(), 15000);
+                    }
+                });
+            else {
+                await message.channel.send({ embeds: [embed], components: actionRows }).then(msg => {
+                    if (react.length > 0) {
+                        const newHandler = (reaction, user) => handler(reaction, user, msg, queue, client, newHandler);
+                        react.forEach(emoji => msg.react(emoji))
+                        client.on('messageReactionAdd', newHandler);
+                    }
+                    else {
+                        if (resetTimeout) {
+                            const timeout = setTimeout(() => msg.delete(), 20000);
+                            client.timeouts[msg.id] = {
+                                timeout,
+                                msg
+                            };
+                        }
+                        setTimeout(() => msg.delete(), 15000);
+                    }
+                });
+            }
         } catch (error) {
             console.error(error);
             const embed = new EmbedBuilder()
